FIX: Various compiler warnings fixed in ZEFoundation. FIX: Various 64bit static code analyzer warnings fixed in ZEFoundation.
svn path=/trunk/; revision=826

+++ Source/ZEFoundation/ZEMath/ZEMath.cpp
@@ -122,5 +122,5 @@ float ZEMath::Lerp(float A, float B, float Factor)
 
 float ZEMath::CopySign(float Value, float Sign)
 {
-	return _copysign(Value, Sign);
+	return (float)_copysign(Value, Sign);
 }
